Allow passing in configure options via semicolon-separated values
This original was thought to be needed to pass in LDFLAGS='option1
option2' on the command line, but it turns out that some (e.g.,
emacs's) configure script trips up over even that (leading me to have
to export the LDFLAGS variable from the build.bash script(s)). But I'm
leaving this code in place just in case it is needed in the future.