emacs build:
Disabled svg_config_options code for now as it does not seem to adversely impact the build.

Add xft as it is needed to use the monofur-13 font in ~/.fonts

Add the system pkg-config path to the PKG_CONFIG_PATH variable so that
system-supplied headers/libraries can be found (this chunk of code
needs to be factored out into a support function).