Remove disabling code
The libtool utility was failing before, but it turned out to be caused
by closing all open file descriptors by a script that was being used
to run the builds in the background.  I don't know why libtool would
fail in this manner or why it needed file descriptor 0 to be left
open, but it is not something I can readily solve.