Add flameshot package
This had to be done because, from what I can gather, the maintainers
do not provide a Ubuntu PPA or anything like that for more recent
releases, and the apt versions seem to be too out of date to be of any
use. And a recent release is needed for my Xubuntu 22.04 installation
as version 11 is now out of date w.r.t. a very egregious defect of:

https://github.com/flameshot-org/flameshot/issues/3232